% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abstract.table}
\alias{abstract.table}
\title{abstract.table}
\format{
An object of class \code{data.frame} with 8 rows and 4 columns.
}
\usage{
abstract.table
}
\description{
Example data to create abstract books
}
\details{
\tabular{llll}{
   title.column \tab authors.column \tab affiliation.column \tab text.column \cr
   title.column \tab authors.column(1) \tab affiliation.column \tab text.column \cr
}
}
\keyword{datasets}
