% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-dict.R
\name{as.Dict}
\alias{as.Dict}
\title{Convert a Data Frame or list to a Dict Object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
as.Dict(x, metadata = NULL, ...)
}
\arguments{
\item{x}{A data frame with columns "name", "label", "labels", "dtype".}

\item{metadata}{A list containing metadata for the Dict object (optional).
It must contain at least a "name" element.
If NULL, the name of the dataset will be used.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A Dict object representing the data dictionary.
}
\description{
Converts a data frame or list representing a data dictionary to a Dict object.
}
\examples{
\dontrun{
    df_dict <- as.data.frame(lbl_df_dict)
    my_dict <- as.Dict(df_dict)
}
}
