% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attendance.table}
\alias{attendance.table}
\title{attendance.table}
\format{
A data frame

\describe{
A data frame with 4 rows and 1 column
}
}
\usage{
attendance.table
}
\description{
Example data to create attendance certificates
}
\keyword{datasets}
