% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{badges.table}
\alias{badges.table}
\title{badges.table}
\format{
A data frame

\describe{
A data frame with 8 rows and 2 columns
}
}
\usage{
badges.table
}
\description{
Example data to create identificative badges
}
\keyword{datasets}
