% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dict.R
\name{browse_dict}
\alias{browse_dict}
\title{Browse Dictionary
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
browse_dict(dict, ...)
}
\arguments{
\item{dict}{A list representing the database dictionary. If not a list, it will be processed by \code{\link{get_dict}}.}

\item{...}{Additional arguments for testing purposes.}
}
\value{
A web interface displaying the data contained in the supplied dictionary.
}
\description{
Allows you to browse the dictionary of the surveys in a web interface, where
you can check the names and labels of the variables, as well as their data labels.
}
\examples{
\dontrun{
dict <- list(
  SEX = list(
    lab = "Sex of the person",
    labs = c("Man" = 1, "Woman" = 2)
  ),
  MARITAL_STATUS = list(
    lab = "Marital status of the person",
    labs = c("Single" = 1, "Married" = 2, "Widower" = 3)
  ),
  HEIGHT = list(
    lab = "Person height",
    labs = c("Short" = 1, "Normal" = 2, "High" = 3)
  )
)
browse_dict(dict)
}
}
