% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{collection.table}
\alias{collection.table}
\title{collection.table}
\format{
An object of class \code{data.frame} with 40 rows and 8 columns.
}
\usage{
collection.table
}
\description{
Example data to create collection labels
}
\details{
\tabular{llllllll}{
   field1 \tab field2 \tab field3 \tab field4 \tab field5 \tab field6 \tab field7 \tab QR_code \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 1 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 01/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 2 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 3 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 4 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Amata huebneri \tab Erebidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 1 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 01/09/1992 \tab Great Yarmouth, Norwich, United Kingdom \tab Item 2 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 3 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 4 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Arion hortensis \tab Arionidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 1 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 01/09/1992 \tab Great Yarmouth, Norwich, United Kingdom \tab Item 2 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 3 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 4 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 1 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 01/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 2 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 13/11/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 3 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 4 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Trorgssionis sp. 1 \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 1 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis sp. 2 \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 2 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis schetelonica \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 1 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis schetelonica \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 2 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 5 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 01/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 6 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 7 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Bombus terrestris subsp. glumbumble \tab Apidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 8 \tab Collected in the coast of the North Sea (it was only known from the Mediterranean Sea before). \tab It produces a fluid that causes melancholy. This fluid is used as an antidote for the hysteria that results from eating Alihotsy leaves. \tab https://static.wikia.nocookie.net/esharrypotter/images/b/b2/HM_Glumblumble.png/revision/latest?cb=20210127125033 \cr
   Amata huebneri \tab Erebidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 5 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 01/09/1992 \tab Great Yarmouth, Norwich, United Kingdom \tab Item 6 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 7 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Amata huebneri \tab Erebidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 8 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Wasp-moth hybrid, giant and aggresive, painful sting \tab https://harrypotter.fandom.com/wiki/Vampyr_Mosp?file=Vampyr_Mosp.png \cr
   Arion hortensis \tab Arionidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 5 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 01/09/1992 \tab Great Yarmouth, Norwich, United Kingdom \tab Item 6 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 13/11/1992 \tab Cargan, Antrim Glens, Northern Ireland. \tab Item 7 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Arion hortensis \tab Arionidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 8 \tab Collected from the only known extant individual, near to the Hogwarts Castle. Damage marks can be seen in the bark of the branches, due to self-harm. \tab Ron Weasley began vomiting slugs after his attempt to curse Draco Malfoy failed due to his broken wand. \tab https://figuritaspop.es/wp-content/uploads/ronweasleyslug1_2.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 05/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 5 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 01/09/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 6 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 13/11/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 7 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Acromantula gigantea \tab Macroaranidae \tab 13/12/1992 \tab Hogwarts Gardens, Hogsmeade, Scotland (UK) \tab Item 8 \tab Forbidden Forest \tab small individuals of approx. 30 cm have been collected. \tab https://mymagicwands.files.wordpress.com/2020/04/img_20200409_115122.jpg \cr
   Trorgssionis sp. 1 \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 5 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis sp. 2 \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 6 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis schetelonica \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 7 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
   Trorgssionis schetelonica \tab Canellae \tab 10/5/23 \tab Grazalema mountain range (Cadiz, Spain) \tab Item 8 \tab Between the bushes of Retama and Nerium \tab Blue colour of the skin indicates adaptation to the field conditions \tab https://www.androidheadlines.com/wp-content/uploads/2021/11/google-search-fishing-troll.jpg \cr
}
}
\keyword{datasets}
