% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accreditation.R
\name{create_accreditation}
\alias{create_accreditation}
\title{Function to create accreditation cards (8 per DIN-A4 page)}
\usage{
create_accreditation(
  data = NULL,
  path = NULL,
  event = NULL,
  name.column = NULL,
  affiliation.column = NULL,
  lpic = NULL,
  rpic = NULL
)
}
\arguments{
\item{data}{a data frame including names and affiliations (optional if \code{affiliation.column} is NULL) to create accreditation cards.}

\item{path}{Folder path where the output will be printed}

\item{event}{Title of the event}

\item{name.column}{Column name of the \code{data} data frame which specifies the participant's name.}

\item{affiliation.column}{Column name of the \code{data} data frame which specifies the participant's affiliation.}

\item{lpic}{PNG object route. File route of the top-left image. Can be blank if set to NULL.}

\item{rpic}{PNG object route. File route of the top-right image. Can be blank if set to NULL.}
}
\value{
An 'output' folder with the PDF documents inside
}
\description{
Function to create accreditation cards (8 per DIN-A4 page)
}
\examples{
data <- read_sheet(url='https://docs.google.com/spreadsheets/d
        /16smXdP-Ehwu1cEmJTbJI1DerIpUrOcD7H5Ni6z9B07M/edit#gid=0')
create_accreditation(data=data,
path = "LabeleR_output",
event="INTERNATIONAL CONFERENCE OF MUGGLEOLOGY",
name.column = "List",
affiliation.column="Affiliation",
lpic = NULL,
rpic = NULL)
}
\author{
Julia G. de Aledo, Ignacio Ramos-Gutierrez
}
