% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate_participation.R
\name{create_certificate_participation}
\alias{create_certificate_participation}
\title{Create certificate of participation}
\usage{
create_certificate_participation(
  language = c("spanish", "english"),
  url = NULL,
  select.column = NULL,
  select.value = NULL,
  type = NULL,
  organiser = NULL,
  hours = NULL,
  signer = NULL,
  signer.position = NULL,
  lpic = NULL,
  rpic = NULL,
  signature.pic = NULL,
  name.column = NULL,
  affiliation.column = NULL,
  date.column = NULL,
  title.column = NULL,
  comm.type.column = NULL
)
}
\arguments{
\item{language}{Select english or spanish}

\item{url}{Specify a valid Google Sheets URL}

\item{select.column}{Column of the Google Sheets that specifies which rows must be selected}

\item{select.value}{Value of \code{select.column} that specifies which rows must be selected}

\item{type}{Type of event (conference, workshop, seminar...)}

\item{organiser}{Name of the organizing entity}

\item{hours}{Number of hours the event has lasted}

\item{signer}{Person who credits the certificate}

\item{signer.position}{Position of the \code{signer}}

\item{lpic}{PNG object. File route of the top-left image. Can be blank if set to NULL.}

\item{rpic}{PNG object. File route of the top-right image. Can be blank if set to NULL.}

\item{signature.pic}{PNG object. File route of a signature image. Can be blank if set to NULL.}

\item{name.column}{Column name of the Google Sheet column which specifies the particpant name.}

\item{affiliation.column}{Column name of the Google Sheet column which specifies the particpant affiliation}

\item{date.column}{Column name of the Google Sheet column which specifies the date.}

\item{title.column}{Column name of the Google Sheet column which specifies the communication title.}

\item{comm.type.column}{Column name of the Google Sheet column which specifies the communication type (oral, poster, online...)}
}
\value{
An 'output' folder with the PDF documents inside
}
\description{
Create certificate of participation
}
\examples{
create_certificate_participation(
language ="en",
url="https://docs.google.com/spreadsheets/u/1/d/11No4aLvta2qxGhkxD7W6HfNfGmO1wpCIDvyRKFF-_gM/edit?usp=drive_web&ouid=106603768357414088091",
select.column=NULL,
select.value=NULL,
type="online seminar",
organiser="Hogwarts School of Witchcraft and Wizardry",
hours=2,
signer="A.P.W.B. Dumbledore",
signer.position="School Headmaster",
lpic="templates/Hogwarts_logo.png",
rpic=NULL,
signature.pic = "templates/firma.png",
name.column="Name",
affiliation.column="House",
date.column="Date",
title.column="Title",
comm.type.column = "Comm.type")

}
\author{
Julia G. de Aledo, Ignacio Ramos-Gutierrez
}
