% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiny_label.R
\name{create_tinylabel}
\alias{create_tinylabel}
\title{Create tiny labels}
\usage{
create_tinylabel(
  data = NULL,
  qr = NULL,
  path = NULL,
  field1.column = NULL,
  field2.column = NULL,
  field3.column = NULL,
  field4.column = NULL,
  field5.column = NULL,
  keep.files = FALSE,
  template = NULL
)
}
\arguments{
\item{data}{a data frame including information of a species}

\item{qr}{String. Free text or column of \code{data} that specifies the link where to create the QR code.
If the specified value is not a column name of \code{data}, all the QRs will be equal,
and will output the specified \code{qr}.}

\item{path}{Character. Path to folder where the PDF file will be saved.}

\item{field1.column}{Character (optional). Name of the column in \code{data} storing the first free text to
appear at the top of the label.}

\item{field2.column}{Character (optional). Name of the column in \code{data} storing the second free text to
appear below field1.}

\item{field3.column}{Character (optional). Name of the column in \code{data} storing the third free text to
appear below field2.}

\item{field4.column}{Character (optional). Name of the column in \code{data} storing the fourth free text to
appear below field3.}

\item{field5.column}{Character (optional). Name of the column in \code{data} storing the fifth free text to
appear below field4.}
}
\value{
A PDF file named "Tiny_label.pdf" is saved on disk, in the folder defined
by \code{path}. If \code{keep.files = TRUE}, an Rmarkdown file will also appear in the same folder.
}
\description{
Create tiny labels (16 per DIN-A4 page)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- read_sheet("https://docs.google.com/spreadsheets/
        d/1Bd_IVgGup4MapTgPq-cqqP05zYl-Q4SfUCBJ5oDSrMs/edit?usp=sharing")
create_tinylabel(
data = tiny.table,
qr = "QR_code",
path = "LabeleR_output",
field1.column = "campo1",
field2.column = "campo2",
field3.column = "campo3",
field4.column = "campo4",
field5.column = "campo5"
)
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutiérrez, Julia G. de Aledo, Francisco Rodríguez-Sánchez
}
