% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-dict.R
\name{db_list_dicts}
\alias{db_list_dicts}
\title{List Dicts in Database
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
db_list_dicts(con)
}
\arguments{
\item{con}{A DBI database connection.}
}
\value{
A data.frame with columns: name, created_at, updated_at.
}
\description{
Lists all dictionary names stored in the database.
}
\examples{
\dontrun{
    con <- DBI::dbConnect(RSQLite::SQLite(), "my_db.sqlite")
    db_list_dicts(con)
    DBI::dbDisconnect(con)
}
}
