% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-dict.R
\name{db_load_dict}
\alias{db_load_dict}
\title{Load a Dict from Database
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
db_load_dict(con, name)
}
\arguments{
\item{con}{A DBI database connection.}

\item{name}{Character string. The name of the dict to load.}
}
\value{
A Dict object.
}
\description{
Loads a Dict object from the database by name.
}
\examples{
\dontrun{
    con <- DBI::dbConnect(RSQLite::SQLite(), "my_db.sqlite")
    my_dict <- db_load_dict(con, "survey_2024")
    DBI::dbDisconnect(con)
}
}
