% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{Dict}
\alias{Dict}
\title{Dict Class
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
Dict(..., metadata)
}
\arguments{
\item{...}{Named lists, each representing a variable in the dictionary.}

\item{metadata}{A list containing metadata about the dictionary itself.}
}
\value{
An object of class "Dict".
}
\description{
Represents a data dictionary containing metadata and variable definitions.
}
\details{
Each variable list should contain the following elements:
\itemize{
\item \strong{label:} A character string representing a short, unique identifier for the variable
and a character string describing the variable. This can be a long, descriptive label.
To provide both a short and a long label, separate them with "||".
For example: "Age || Age of the respondent".
Use \code{set_Dict(label_type=)} argument to choose the desired label type.
By default, the full label (including both parts) will be used.
\item \strong{labels:} (optional) A named vector or list defining the value labels for the variable.
\item \strong{dtype:} (optional) A character string specifying the desired data type of the variable.
}

The \strong{\code{metadata}} list can contain any relevant information about the dictionary,
such as the source of the data, creation date, etc. It must contain at least a "name" element.
}
\examples{
\dontrun{
    lbl_df_dict |> unclass()
    is.Dict(lbl_df_dict)
    lbl_df_dict
}
}
