% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{dict_browser}
\alias{dict_browser}
\title{Browse a Data Dictionary Interactively
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
dict_browser(x, .interactive = interactive())
}
\arguments{
\item{x}{A \code{Dict} object or a labeled \code{data.frame}.}

\item{.interactive}{Logical. If TRUE (default and when interactive session is detected),
an interactive DT::datatable is displayed. If FALSE, a data frame representation of
the dictionary is returned.}
}
\value{
An interactive DT::datatable (if \code{.interactive} is TRUE) or a data frame
representation of the dictionary (if \code{.interactive} is FALSE).
}
\description{
Displays a data dictionary in an interactive web interface using the DT package.
Accepts either a \code{Dict} object or a labeled \code{data.frame} as input.
}
\examples{
\dontrun{
    # Passing a Dict object
    dict_browser(lbl_df_dict)
    # Passing a labeled data.frame will automatically generate a Dict object
    dict_browser(set_Dict(lbl_df, lbl_df_dict))
    
    # For non-interactive use:
    df <- dict_browser(lbl_df_dict, .interactive = FALSE)
    print(df)
}
}
