% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{email_configuration}
\alias{email_configuration}
\title{Setup the email server}
\usage{
email_configuration(
  user,
  app.name = NULL,
  subject = NULL,
  body = NULL,
  cc = NULL,
  bcc = NULL
)
}
\arguments{
\item{user}{Character. Email account used to send certificates.}

\item{app.name}{Name of the mail application used to send emails. To create one, access \verb{https://myaccount.google.com/apppasswords}}

\item{subject}{Character. Subject of the email to be sent If not specified, labeleR will use a default value.}

\item{body}{Character. Body text of the email to be sent. If not specified, labeleR will use a default value.}

\item{cc}{Character. String (or vector of strings) containing the email addresses to send the email as a copy.}

\item{bcc}{Character. String (or vector of strings) containing the email addresses to send the email as a hidden copy.}
}
\value{
A list including at least a 'user' string' and an 'app.name' string. Optionally, slots 'subject',
'body', 'cc' and 'bcc' can be edited to compile the email to send.
}
\description{
Create the email.info object to feed the certificate creation functions
}
\author{
Ignacio Ramos-Gutierrez, Julia G. de Aledo, Jimena Mateo-Martín, Francisco Rodriguez-Sanchez
}
