% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{Example dataset}
\format{
\subsection{\code{exampledata}}{

A data frame with 30 rows and 2 columns:
\describe{
\item{x}{Random numbers between 0 and 1}
\item{y}{Random numbers between 0 and 1}
}
}
}
\source{
dataset simulated
}
\usage{
exampledata
}
\description{
This is just an example dataset
}
\keyword{datasets}
