% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json-dict.R
\name{from_json}
\alias{from_json}
\title{Read a Dict Object from JSON
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
from_json(x)
}
\arguments{
\item{x}{Either a JSON string or a file path to a JSON file.}
}
\value{
A Dict object.
}
\description{
Reads a JSON string or file and converts it to a Dict object.
}
\examples{
\dontrun{
    # From file
    my_dict <- from_json("my_dictionary.json")

    # From string
    json_str <- '{"metadata": {"name": "Test"}, "variables": {"x": {"label": "X"}}}'
    my_dict <- from_json(json_str)
}
}
