% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_dict.R
\name{get_dict}
\alias{get_dict}
\title{Generate a dictionary of given dataset
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
get_dict(tbl)
}
\arguments{
\item{tbl}{A labelled dataset}
}
\value{
A dictionary of the dataset
}
\description{
Generate a dictionary of given dataset
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
 get_dict(tbl)
}
}
