% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbarium.table}
\alias{herbarium.table}
\title{herbarium.table}
\format{
A data frame containing information on taxonomy, collection details, etc.
}
\usage{
herbarium.table
}
\description{
Example data to create herbarium labels
}
\keyword{datasets}
