% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is-dict.R
\name{is.Dict}
\alias{is.Dict}
\title{Is Method for Dict Objects
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
is.Dict(x)
}
\arguments{
\item{x}{The object to check.}
}
\value{
TRUE if x is a valid Dict object, FALSE otherwise.
}
\description{
Is Method for Dict Objects
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
is.Dict(lbl_df_dict)
}
}
