% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{label.table}
\alias{label.table}
\title{label.table}
\format{
A data frame containing information on collected specimens.
}
\usage{
label.table
}
\description{
Example data to create collection labels
}
\keyword{datasets}
