% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeleR-package.R
\docType{package}
\name{labeleR-package}
\alias{labeleR}
\alias{labeleR-package}
\title{labeleR: Package to create your own labels, certificates, and much more! :)}
\description{
With this packages you can create labels, herbarium labels, ID cards, custom certificates
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/labeleR}
  \item Report bugs at \url{https://github.com/EcologyR/labeleR/issues}
}

}
\author{
\strong{Maintainer}: Ignacio Ramos-Gutiérrez \email{ig.ramosgutierrez@gmail.com} (\href{https://orcid.org/0000-0002-8675-0114}{ORCID})

Authors:
\itemize{
  \item Julia G. de Aledo \email{juliagdealedo@gmail.com} (\href{https://orcid.org/0000-0001-9065-9316}{ORCID})
  \item Francisco Rodríguez-Sánchez \email{frodriguez@us.es} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
}

}
\keyword{internal}
