% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{lbl_df}
\alias{lbl_df}
\title{Example Labeled Dataset}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{sex}{Numeric, representing sex (1 = Male, 2 = Female).}
\item{age}{Numeric, representing age.}
}
}
\source{
Simulated data.
}
\usage{
lbl_df
}
\description{
A simple data frame with two variables: sex and age.
}
\keyword{datasets}
