% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{lbl_df_dict}
\alias{lbl_df_dict}
\title{Example Data Dictionary for lbl_df}
\format{
A Dict object with the following variable definitions:
\describe{
\item{sex}{
\itemize{
\item label: "Sex || What is your sex?"
\item labels: c("Male" = 1, "Female" = 2)
}
}
\item{age}{
\itemize{
\item label: "Age || How old are you?"
\item dtype: "integer"
}
}
}
}
\source{
Simulated data.
}
\usage{
lbl_df_dict
}
\description{
A Dict object defining the data dictionary for the \code{lbl_df} dataset.
}
\keyword{datasets}
