% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_scatterplot.R
\name{make_scatterplot}
\alias{make_scatterplot}
\title{Make a scatterplot of two numeric variables}
\usage{
make_scatterplot(var1 = NULL, var2 = NULL, ...)
}
\arguments{
\item{var1}{Numeric vector}

\item{var2}{Numeric vector}

\item{...}{Optional arguments to be passed to \code{\link[ggplot2:labs]{ggplot2::labs()}} such as \code{title},
\code{x} or \code{y} to define plot title, horizontal axis or vertical axis names,
respectively}
}
\description{
Make a scatterplot of two numeric variables using \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\examples{
data("exampledata")
make_scatterplot(exampledata$x, exampledata$y, title = "My scatterplot")
}
