% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multichoice.table}
\alias{multichoice.table}
\title{multichoice.table}
\format{
An object of class \code{data.frame} with 7 rows and 5 columns.
}
\usage{
multichoice.table
}
\description{
Example data to create multiple coice tests
}
\details{
\tabular{lllll}{
   question \tab opt1.correct \tab opt2 \tab opt3 \tab opt4 \cr
   In which Hogwarts house was Harry sorted in? \tab Gryffindor \tab Slytherin \tab Hufflepuff \tab Revenclaw \cr
   Who was Dobby? \tab A house elf \tab A hippogryph \tab A muggle \tab A Hogwarts teacher \cr
   Which was Longbottom's name? \tab Neville \tab Ronald \tab Bill \tab Luna \cr
   Which was Hogwarts' most famous tree? \tab The Whomping Willow \tab The Elder Alder \tab The Mighty Beech \tab The Robust Oak \cr
   Which is the nearest town to Hogwarts? \tab Hogsmeade \tab London \tab Godric's Hollow \tab Little Hangleton \cr
   Who was Harry's Godfather? \tab Sirius Black \tab Remus Lupin \tab Severus Snape \tab Lucius Malfoy \cr
   Which of the following is not a middle name from Albus Dubledore: \tab Billius \tab Wulfric \tab Brian \tab Percival \cr
}
}
\keyword{datasets}
