% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{participation.table}
\alias{participation.table}
\title{participation.table}
\format{
A data frame

\describe{
A data frame with 2 rows and 5 columns
}
}
\usage{
participation.table
}
\description{
Example data to create participation certificates
}
\keyword{datasets}
