% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sheet.R
\name{read_sheet}
\alias{read_sheet}
\title{Title}
\usage{
read_sheet(url, select.column = NULL, select.value = NULL)
}
\arguments{
\item{url}{a valid google sheets url including a table}

\item{select.column}{column name to filter which rows to include}

\item{select.value}{if select.column is not NULL, the value of the rows to be selected}
}
\value{
a data frame
}
\description{
Title
}
\examples{
table <- read_sheet(url='https://docs.google.com/spreadsheets/d/1inkk3_oNvvt8ajdK4wOkSgPoUyE8JzENrZgSTFJEFBw/edit#gid=0')
}
