% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resta.R
\name{resta}
\alias{resta}
\title{Subtract numeric values}
\usage{
resta(x = NULL, y = NULL, warn.negative = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector. Values to subtract from x}

\item{warn.negative}{Logical Warn if result is negative}
}
\value{
A number or numeric vector of the same length as x
}
\description{
Subtract numeric values
}
\examples{
resta(2, 3)
resta(c(2, 2), c(3, 3))

\dontrun{
resta(c(2, 2), c(3, 3), warn.negative = TRUE)  # gives warning
}
}
