% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{set_Dict}
\alias{set_Dict}
\title{Apply a Dictionary to a Dataset
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
set_Dict(
  tbl,
  dict,
  subset = NULL,
  dtypes = TRUE,
  label_type = c("full", "label", "name")
)
}
\arguments{
\item{tbl}{A data.frame or tibble.}

\item{dict}{A Dict object.}

\item{subset}{(Optional) A character vector specifying the variables to which the dictionary should be applied.
If NULL (default), the dictionary is applied to all variables in the dataset that are also present in the dictionary.}

\item{dtypes}{Logical, indicating whether to apply data types defined in the dictionary (default is TRUE).}

\item{label_type}{Character string specifying how variable labels should be applied.
Options are "full" (default, uses the entire label), "label" (uses the part after "||"), or "name" (uses the part before "||").
See \code{Dict} for more details.}
}
\value{
The dataset with labels and data types applied according to the dictionary.
}
\description{
Applies labels and data types defined in a Dict object to a dataset.
}
\examples{
\dontrun{
    labeled_data <- set_Dict(lbl_df, lbl_df_dict)
}
}
