% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labels.R
\name{set_labels}
\alias{set_labels}
\title{Assign data labels to specified variables
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
set_labels(tbl, dict, vars = NULL, ignore_case = FALSE, warn = TRUE)
}
\arguments{
\item{tbl}{\link{data.frame}: Data.frame with the data}

\item{dict}{\link{data.frame}: Dictionary with all the data labels to use}

\item{vars}{\link{character}: If specified, only labels are assigned to those variables}

\item{ignore_case}{\link{logical}: Indicate if case sensitive should be ignored.}

\item{warn}{\link{logical}: Indicate if warnings should be shown.}
}
\value{
The data entered in the \code{tbl} argument but with data labels
}
\description{
Assign data labels to specified variables
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
\dontrun{
dict <- list(
  SEX = list(
    lab = "Sex of the person",
    labs = c(
      "Man" = 1,
      "Woman" = 2
    )
  )
)
df <- data.frame(SEX = c(1, 2))
str(df)
str(set_labels(df, dict = dict))
}
}
\seealso{
Data labels \code{vignette("labeler", package = "labeler")}
}
