% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suma.R
\name{suma}
\alias{suma}
\title{Sum numeric values}
\usage{
suma(x = NULL, y = NULL)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric}
}
\value{
A number or numeric vector
}
\description{
This function sums numeric vectors
}
\examples{
suma(2, 3)
suma(c(2, 2), c(3, 3))
}
