% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateRpackage-package.R
\docType{package}
\name{templateRpackage-package}
\alias{templateRpackage}
\alias{templateRpackage-package}
\title{templateRpackage: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/templateRpackage}
  \item Report bugs at \url{https://github.com/EcologyR/templateRpackage/issues}
}

}
\author{
\strong{Maintainer}: Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})

}
\keyword{internal}
