% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tiny.table}
\alias{tiny.table}
\title{tiny.table}
\format{
A data frame

\describe{
A data frame with 40 rows and 8 columns
}
}
\usage{
tiny.table
}
\description{
Example data to create tinylabels
}
\keyword{datasets}
