% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{with_Dict}
\alias{with_Dict}
\title{Work with a Dataset Using a Dictionary
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
with_Dict(
  tbl,
  dict = NULL,
  subset = NULL,
  use_label = TRUE,
  use_labels = TRUE,
  ...
)
}
\arguments{
\item{tbl}{A data.frame or tibble.}

\item{dict}{A Dict object (optional). If NULL, it assumes the dataset already has a dictionary applied.}

\item{subset}{(Optional) A character vector specifying variables to work with.}

\item{use_label}{Logical, indicating whether to replace column names with variable label (default is TRUE).}

\item{use_labels}{Logical, indicating whether to convert variables to factors with labels as levels (default is TRUE).}

\item{...}{Additional arguments passed to \code{set_Dict}.}
}
\value{
The modified dataset.
}
\description{
A convenience function to apply a dictionary to a dataset and optionally convert variables to factors
with labels as levels or replace column names with variable labels.
}
\examples{
\dontrun{
    factored_data <- with_Dict(lbl_df, lbl_df_dict, use_labels = TRUE)
    renamed_data <- with_Dict(lbl_df, lbl_df_dict, use_label = TRUE)
}
}
