% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{as.Dict}
\alias{as.Dict}
\title{Convert a Data Frame to a Dict Object
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
as.Dict(x, metadata = NULL, ...)
}
\arguments{
\item{x}{A data frame with columns "name", "label", "type", "labels".}

\item{metadata}{A list containing metadata for the Dict object (optional).}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A Dict object representing the data dictionary.
}
\description{
Converts a data frame representing a data dictionary to a Dict object.
}
\examples{
# Assuming 'df' is a data frame with appropriate columns
my_dict <- as.Dict(df)
}
