% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attendance.table}
\alias{attendance.table}
\title{attendance.table}
\format{
An object of class \code{data.frame} with 4 rows and 1 columns.
}
\usage{
attendance.table
}
\description{
Example data to create attendance certificates
}
\details{
A data frame with the name of attendees\tabular{l}{
   Names \cr
   Harry Potter \cr
   Hermione Granger \cr
   Neville Longbottom \cr
   Ginny Weasley \cr
}
}
\keyword{datasets}
