% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abstractbook.R
\name{create_abstractbook}
\alias{create_abstractbook}
\title{Create Book of Abstracts}
\usage{
create_abstractbook(
  data = NULL,
  path = NULL,
  filename = NULL,
  title.column = NULL,
  authors.column = NULL,
  affiliation.column = NULL,
  affiliation.address.column = NULL,
  text.column = NULL,
  frontpage = NULL,
  title.cex = 22,
  authors.cex = 16,
  affiliations.cex = 14,
  text.cex = 14,
  keep.files = FALSE,
  template = NULL
)
}
\arguments{
\item{data}{a data frame including abstract titles, names, affiliations and abstract texts.}

\item{path}{Character. Path to folder where the PDF file will be saved.}

\item{filename}{Character. Filename of the pdf. If NULL, default is "AbstractBook".}

\item{title.column}{Character. Name of the column in \code{data}
storing abstracts' titles.}

\item{authors.column}{Character. Name of the column in \code{data}
storing abstracts' names. Different authors names must be separated using a semi-colon (';').}

\item{affiliation.column}{Character (optional). Name of the column in \code{data} storing the numbers
or symbols (e.g. '*' for corresponding author) representing each autor affiliations.
Different affiliations of each author  must be specified using a comma (',') as separator.
Separations between each author's affiliations must be specified using a semi-colon (';').}

\item{affiliation.address.column}{Name of the column in \code{data} specifying the address for
each affiliation. The number shoud be the same followed in \code{affiliation.column}.}

\item{frontpage}{Character. Path to PDF file to be inserted before the book of abstracts (as front page and/or introduction).}

\item{title.cex}{Text font size used for the title. Default is 22.}

\item{authors.cex}{Text font size used for the authors' names. Default is 16.}

\item{affiliations.cex}{Text font size used for the affiliation addresses Default is 14.}

\item{text.cex}{Text font size used for the abstract main text body. Default is 14.}

\item{keep.files}{Logical. Keep the RMarkdown template and associated files
in the output folder? Default is FALSE.}

\item{template}{Character (optional) RMarkdown template to use. If not provided,
using the default template included in \code{labeleR}.}

\item{text.colum}{Name of the column in \code{data} storing the abstract text.}
}
\value{
A PDF file named after \code{filename} is saved on disk, in the folder defined
by \code{path}. If \code{keep.files = TRUE}, an RMarkdown will also
appear in the same folder.
}
\description{
Create a book of abstracts given a dataframe or tibble.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
\dontshow{\}) # examplesIf}
}
\author{
Ignacio Ramos-Gutierrez, Julia G. de Aledo, Francisco Rodriguez-Sanchez
}
