% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certificate_attendance.R
\name{create_certificate_attendance}
\alias{create_certificate_attendance}
\title{Create certificate of attendance}
\usage{
create_certificate_attendance(
  data = NULL,
  path = NULL,
  language = "english",
  type = NULL,
  title = NULL,
  organiser = NULL,
  hours = NULL,
  signer = NULL,
  signer.position = NULL,
  date = NULL,
  speaker = NULL,
  lpic = NULL,
  rpic = NULL,
  signature.pic = NULL,
  name.column = NULL
)
}
\arguments{
\item{data}{a data frame  to create attendance certificates.}

\item{path}{Folder path where the output will be printed}

\item{language}{Select english or spanish}

\item{type}{Type of event (conference, workshop, seminar...)}

\item{title}{Title of the event}

\item{organiser}{Name of the organizing entity}

\item{hours}{Number of hours the event has lasted}

\item{signer}{Person who credits the certificate}

\item{signer.position}{Position of the \code{signer}}

\item{date}{Date of the event}

\item{speaker}{Name of the speaker of the event}

\item{lpic}{PNG object. File route of the top-left image. Can be blank if set to NULL.}

\item{rpic}{PNG object. File route of the top-right image. Can be blank if set to NULL.}

\item{signature.pic}{PNG object. File route of a signature image. Can be blank if set to NULL.}

\item{name.column}{Column name of the Google Sheet column which specifies the attendee's name.}
}
\value{
An 'output' folder with the PDF documents inside
}
\description{
Create certificate of attendance
}
\examples{
\dontrun{
create_certificate_attendance(
data= read_sheet('https://docs.google.com/spreadsheets/
        d/1inkk3_oNvvt8ajdK4wOkSgPoUyE8JzENrZgSTFJEFBw/edit#gid=0'),
path = "LabeleR_output",
language="en",
type="class",
title="Potions Class",
organiser="Hogwarts School year 1992-1993",
signer="A.P.W.B. Dumbledore",
signer.position="School Headmaster",
hours=200,
date="01/01/2021",
speaker="Severus Snape",
rpic=NULL,
lpic=NULL,
signature.pic=NULL,
name.column="List_assistants"
)
}
}
\author{
Julia G. de Aledo, Ignacio Ramos-Gutierrez
}
