% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collection_label.R
\name{create_collection_label}
\alias{create_collection_label}
\title{Function to create create collection labels (8 per DIN-A4 page)}
\usage{
create_collection_label(
  data = data,
  path = NULL,
  qr = NULL,
  field1.column = NULL,
  field2.column = NULL,
  field3.column = NULL,
  field4.column = NULL,
  field5.column = NULL,
  logo = NULL,
  bgcolor = NULL,
  textcolor = NULL
)
}
\arguments{
\item{data}{Data frame including to create labels.}

\item{path}{Folder path where the output will be printed}

\item{qr}{String. Free text or column of \code{data} that specifies the text to create the QR code.
If the specified value is not a column name of \code{data}, all the QRs will be equal, and will output the
specified \code{qr}.}

\item{field1.column}{Column of \code{data} that specifies the text to print in the first field.}

\item{field2.column}{Column of \code{data} that specifies the text to print in the second field.}

\item{field3.column}{Column of \code{data} that specifies the text to print in the third field.}

\item{field4.column}{Column of \code{data} that specifies the text to print in the fourth field.}

\item{field5.column}{Column of \code{data} that specifies the text to print in the fifth field.}

\item{logo}{PNG object route. File route of the bottom image. Can be blank if set to NULL.}

\item{bgcolor}{HTML color to use for the label background. Default is D0ECC1}

\item{textcolor}{HTML color to use for the label text. Default is 1E3F20}
}
\description{
Function to create create collection labels (8 per DIN-A4 page)
}
\examples{
data <- read_sheet("https://docs.google.com/spreadsheets/d/
        1Bd_IVgGup4MapTgPq-cqqP05zYl-Q4SfUCBJ5oDSrMs/edit?usp=sharing")

create_collection_label(
data = data,
path = "LabeleR_output",
qr = "QR_code",
field1.column = "field1",
field2.column = "field2",
field3.column = "field3",
field4.column = "field6",
field5.column = "field7"

)


}
\author{
Julia G. de Aledo, Ignacio Ramos-Gutierrez
}
