% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/herbarium.R
\name{create_herbarium_label}
\alias{create_herbarium_label}
\title{Function to create accreditation cards in DIN-A7 size}
\usage{
create_herbarium_label(
  data = data,
  title = NULL,
  subtitle = NULL,
  family.column = NULL,
  taxon.column = NULL,
  author.column = NULL,
  det.column = NULL,
  date.det.column = NULL,
  location.column = NULL,
  area.description.column = NULL,
  latitude.column = NULL,
  longitude.column = NULL,
  elevation.column = NULL,
  field1.column = NULL,
  field2.column = NULL,
  field3.column = NULL,
  collector.column = NULL,
  collection.column = NULL,
  assistants.column = NULL,
  date.column = NULL
)
}
\arguments{
\item{title}{Main title at the top of the labels. Can be blank if set to NULL.}

\item{subtitle}{Subtitle at the bottom of the labels. Can be blank if set to NULL.}

\item{family.column}{Column name of the \code{data} data frame which specifies the labels' family name.}

\item{taxon.column}{Column name of the \code{data} data frame which specifies the labels' taxon.}

\item{author.column}{Column name of the \code{data} data frame which specifies the taxons' author.}

\item{det.column}{Column name of the \code{data} data frame which specifies the determiner of the voucher.}

\item{date.det.column}{Column name of the \code{data} data frame which specifies the date when the voucher was determined.}

\item{area.description.column}{Column name of the \code{data} data frame which specifies the decription of the area}

\item{latitude.column}{Column name of the \code{data} data frame which specifies the latitude where the specimen was collected.}

\item{elevation.column}{Column name of the \code{data} data frame which specifies the elevation where the specimen was collected.}

\item{field1.column}{Column name of the \code{data} data frame which specifying a variable of the user's choice. Can be blank if set to NULL.}

\item{field2.column}{Column name of the \code{data} data frame which specifying a variable of the user's choice. Can be blank if set to NULL.}

\item{field3.column}{Column name of the \code{data} data frame which specifying a variable of the user's choice. Can be blank if set to NULL.}

\item{collector.column}{Column name of the \code{data} data frame which specifies the name of the collector of the voucher.}

\item{assistants.column}{Column name of the \code{data} data frame which specifies the names of the collector's assistants.}

\item{date.column}{Column name of the \code{data} data frame which specifies the date when the specimen was collected.}

\item{longitude.columnColumn}{name of the \code{data} data frame which specifies the longitude where the specimen was collected.}
}
\value{
A pdf file with four herbarium labels per page within an 'output' folder
}
\description{
Function to create accreditation cards in DIN-A7 size
}
\examples{
#
data=read_sheet("https://docs.google.com/spreadsheets/d/1Q005BDM0XyUNq5XzGWuvdzgZVMc4KhbYadVzi77h3Xw/edit?usp=sharing")
create_herbarium_label(
data=data,
title="Magical flora of the British Isles",
subtitle="Project: Eliminating plant blindness in Hogwarts students",
family.column="Family",
taxon.column="Taxon",
author.column="Author",
det.column="det/conf",
date.det.column="Det_date",
location.column="Location",
area.description.column="Area_description",
latitude.column="Latitude",
longitude.column="Longitude",
elevation.column="Elevation",
field1.column="life_form",
field2.column="Observations",
field3.column="Height",
collector.column="Collector",
collection.column="Collection_number",
assistants.column="Assistants",
date.column="Date"
)
}
\author{
Julia G. de Aledo, Ignacio Ramos-Gutierrez
}
