% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict.R
\name{gen_Dict}
\alias{gen_Dict}
\title{Generate a Dictionary from a Dataset
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
gen_Dict(tbl, metadata = attributes(tbl)$dict_metadata)
}
\arguments{
\item{tbl}{A labeled dataset.}

\item{metadata}{A list containing metadata for the dictionary.
It must contain at least a "name" element.
If NULL, the name of the dataset will be used.}
}
\value{
A Dict object representing the dictionary of the dataset.
}
\description{
Creates a Dict object from a labeled dataset (data.frame or tibble).
}
\examples{
\dontrun{
    my_dict <- gen_Dict(lbl_df)
}
}
