% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{herbarium.table}
\alias{herbarium.table}
\title{herbarium.table}
\format{
A data frame

\describe{
A data frame with 6 rows and 19 columns
}
}
\usage{
herbarium.table
}
\description{
Example data to create herbarium labels
}
\keyword{datasets}
