% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeleR-package.R
\docType{package}
\name{labeleR-package}
\alias{labeleR}
\alias{labeleR-package}
\title{labeleR: Automate the Production of Custom Labels, Badges, Certificates, and Other Documents}
\description{
Create custom labels, badges, certificates and other documents. Automate the production of potentially large numbers of herbarium and collection labels, accreditation badges, attendance and participation certificates, etc, and deliver them automatically. Documents are generated in PDF format, which requires a working installation of 'LaTeX', such as 'TinyTeX'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/labeleR/}
  \item \url{https://ecologyr.github.io/labeleR/}
  \item Report bugs at \url{https://github.com/EcologyR/labeleR/issues/}
}

}
\author{
\strong{Maintainer}: Ignacio Ramos-Gutierrez \email{ig.ramosgutierrez@gmail.com} (\href{https://orcid.org/0000-0002-8675-0114}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Julia G. de Aledo \email{juliagdealedo@gmail.com} (\href{https://orcid.org/0000-0001-9065-9316}{ORCID}) [copyright holder]
  \item Jimena Mateo-Martín \email{jimena.mateo.martin@gmail.com} (\href{https://orcid.org/0000-0002-9757-5497}{ORCID}) [copyright holder]
  \item Francisco Rodriguez-Sanchez \email{f.rodriguez.sanc@gmail.com} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Giorgio Comai [contributor]
}

}
\keyword{internal}
