% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeleR-package.R
\docType{package}
\name{labeleR-package}
\alias{labeleR}
\alias{labeleR-package}
\title{labeleR: Automate the Production of Custom Labels, Badges, Certificates, and Other Documents}
\description{
'labeleR' makes it easy to create custom labels, badges, certificates and other documents. It can be used to automate the production of potentially large numbers of accreditation badges, attendance and participation certificates, herbarium and collection labels, etc. Documents are generated in PDF format, which requires a working installation of 'LaTeX', such as 'TinyTeX'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/labeleR}
  \item Report bugs at \url{https://github.com/EcologyR/labeleR/issues}
}

}
\author{
\strong{Maintainer}: Ignacio Ramos-Gutiérrez \email{ig.ramosgutierrez@gmail.com} (\href{https://orcid.org/0000-0002-8675-0114}{ORCID})

Authors:
\itemize{
  \item Julia G. de Aledo \email{juliagdealedo@gmail.com} (\href{https://orcid.org/0000-0001-9065-9316}{ORCID})
  \item Francisco Rodríguez-Sánchez \email{frodriguez@us.es} (\href{https://orcid.org/0000-0002-7981-1599}{ORCID})
}

}
\keyword{internal}
