% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_dict.R
\name{parse_dict}
\alias{parse_dict}
\title{Replace non ASCII characters with ASCII equivalents}
\usage{
parse_dict(dict)
}
\arguments{
\item{dict}{A dictionary. See vignette('labeler') for details.}
}
\value{
A dictionary with all non ASCII characters replaced with ASCII
}
\description{
Replace non ASCII characters with ASCII equivalents
}
\examples{
\dontrun{
   parse_dict(dict)
 }
}
